# PTRef: Some perf improvements for /line_reports