--- a/source/ptreferential/ptref_graph.cpp
+++ b/source/ptreferential/ptref_graph.cpp
@@ -88,9 +88,11 @@ Jointures::Jointures() {
     boost::add_edge(vertex_map.at(Type_e::Line), vertex_map.at(Type_e::Route), g);
     boost::add_edge(vertex_map.at(Type_e::VehicleJourney), vertex_map.at(Type_e::Route), g);
 
-    // Avoid JourneyPattern-JourneyPatternPoint on Route <-> StopPoint for performance issue
+    // Avoid JourneyPattern-JourneyPatternPoint on Route <-> StopPoint and Route <-> StopArea for performance issue
     boost::add_edge(vertex_map.at(Type_e::StopPoint), vertex_map.at(Type_e::Route), Edge(2.5), g);
     boost::add_edge(vertex_map.at(Type_e::Route), vertex_map.at(Type_e::StopPoint), Edge(2.5), g);
+    boost::add_edge(vertex_map.at(Type_e::StopArea), vertex_map.at(Type_e::Route), Edge(3.3), g);
+    boost::add_edge(vertex_map.at(Type_e::Route), vertex_map.at(Type_e::StopArea), Edge(3.3), g);
 
     // JourneyPattern => {Route, JourneyPatternPoint, VehicleJourney, PhysicalMode}
     boost::add_edge(vertex_map.at(Type_e::Route), vertex_map.at(Type_e::JourneyPattern), g);