--- a/source/type/stop_point.h
+++ b/source/type/stop_point.h
@@ -34,6 +34,7 @@ www.navitia.io
 #include "type/geographical_coord.h"
 #include <vector>
 #include <set>
+#include <boost/container/flat_set.hpp>
 #include "type/fwd_type.h"
 
 namespace navitia {
@@ -54,7 +55,7 @@ struct StopPoint : public Header, Nameable, hasProperties, HasMessages {
     Network* network;
     std::vector<StopPointConnection*> stop_point_connection_list;
     std::set<Dataset*> dataset_list;
-    std::set<Route*> route_list;
+    boost::container::flat_set<Route*> route_list;
 
     template <class Archive>
     void serialize(Archive& ar, const unsigned int);