Different optimizations on some bottlenecks of requests like `commercial_modes/<toto>/line_reports` and `networks/<toto>/line_reports`.

:mag: please see commits with message for details, as each should be independent.

:heart: for upcoming thoughts like:  
:speaking_head: "How you optimized for this case (say 'insert') may slow this other case (say 'bina/load' or 'small requests')"


### Perf issues left

Local perfs (no disruptions):
`coverage/TCL/networks/TCL/l_r` : 1.4 s
`coverage/TN/networks/RATPBus/l_r` : 2.6 s
`coverage/IDF/c_m/Bus/l_r` : 1 min 7 s
`coverage/TN/c_m/Bus/l_r` : 1 min 35 s


### Ideas left to explore

* complete tests on pathes in PTRef-graph : https://jira.kisio.org/browse/NAVP-1404

On perfs (https://jira.kisio.org/browse/NAVP-1404) :
1. Compare/audit `add_objects` requests (perfs also) versus corresponding hand-made requests (networks/toto/stop_areas + networks/toto/stop_points + networks/toto/routes + networks/toto/networks... beware line is used along the way to sort objects).
2. Stop iterating on lines in PTRef, just sort after
3. Re-use collections built for the next one when relevant (re-use `routes` for `stop_points`)
4. Request "the other way around"? Ask for all the `Impacts` then sort them? Or do some PTRef tricks from there?
  > Maybe store a "connecting-factor" along the links to anticipate the final collection... And if this results in a figure more important than the collection itself, it may be worth iterating backward. (we can even add a more precise first factor by using the actual size of the collection used).
5. Replace `std::Set` by `boost::flat_set` everywhere in data types for contiguity
6. Replace `std::Vector` by `boost::flat_set` everywhere in data types as there is no duplicates
7. Expand the use of `std::Set<Idx>` to replace `Indexes`(flat_set) in PTRef

Bonus ideas after discussion with @pbench :

8. Implicit iterations over intermediary collections (never create those std::Set that just replaced flat_set)
9. Build a vector instead of a set and care about uniqueness only at the end. Might be dangerous when collections decrease (ex: request the networks of a lot of StopAreas)
10. Avoid the creation of Indexes from any initial data link collection. Maybe by providing a collection to be filled (side-effect :man_shrugging: )

### Performance overview along the way

Local perfs (no disruptions)

| kraken version | RAM (TN) | `networks/` `TCL/l_r` | `networks/` `RATPBus/l_r` | `coverage/TN` `/c_m/Bus/l_r` |
|--|--|--|--|--|
| No optim/shortcut (not even Route-SP)| 1934M | 37 s | 88 s | 2 h 35 min |
| Route-SP shortcut after #2946 and #2931 | +0.75% | 6.5 s | 15.5 s | 19 min |
| Route-SP + Route-SA shortcuts | +1.5% | ? | 4.8 s | ? |
| Only std::Set in get_target_by_source() | = | 4.5 s | 6.8 s | 3 min 8 s |
| Route-SP + Route-SA + Set | +1.5% | ? | 3.7 s | ? |
| All this PR (flat_set + alloc) | -0.4% | 1.4 s | 2.6 s | 1 min 35 s |


----


JIRA: https://jira.kisio.org/browse/NAVITIAII-2793
Following work on #2946 and #2931 .